+++ src/bin/statinamic-setup.js
@@ -1,38 +1,74 @@
-import debug from "debug"
-import npmInstallPackage from "npm-install-package"
+import "babel-polyfill"
 
+import program from "commander"
+import color from "chalk"
+import { join } from "path"
+import fs from "fs-extra"
+
+import { prompt } from  "./utils/inquirer"
+import questions, { defaultTestAnswers } from "./data/questions"
+import template from "./data/template"
 import {
+  version as statinamicVersion,
   peerDependencies as peerDeps,
   optionalPeerDependencies as opPeerDeps,
 } from "../../package.json"
 
-const tobeInstalledDeps = {
-  ...peerDeps,
-  ...opPeerDeps,
-}
+// Currently, we have to wrap this in a function
+// to ensure that babel-polyfill is placed in the top
+// TODO: Take a look at this https://github.com/marten-de-vries/kneden
+(function() {
+  program
+    .option("-t, --test", "Test mode. Default value, no statinamic in devDeps")
+    .parse(process.argv);
 
-const log = debug("statinamic:bin:setup")
+  const cwd = process.cwd()
 
-log("Installing required dependencies...")
+  async function main() {
+    try {
+      let answers
+      const testMode = program.test
 
-const depsMap = Object.keys(tobeInstalledDeps)
-  .reduce(
-    (deps, dep) => {
-      deps.push(`${ dep }@${ tobeInstalledDeps[dep] }`)
-      return deps
-    },
-    []
-  )
-npmInstallPackage(
-  depsMap,
-  {
-    saveDev: true,
-  },
-  (err) => {
-    if (err) {
-      throw err
+      if (testMode) {
+        answers = defaultTestAnswers
+      }
+      else {
+        answers = await prompt(questions)
+      }
+      const { name, homepage, ...statinamic } = answers
+
+      const devDependencies = {
+        ...peerDeps,
+        ...opPeerDeps,
+        ...!testMode && {
+          statinamic: `^${ statinamicVersion }`,
+        },
+      }
+
+      const pkg = {
+        ...template,
+        name,
+        homepage,
+        statinamic,
+        devDependencies,
+      }
+
+      fs.writeJsonSync(join(cwd, "package.json"), pkg)
+      console.log(color.green("Generated package.json file"))
+
+      const boilerplatePath = join(__dirname, "../../boilerplate")
+      fs.copySync(boilerplatePath, cwd)
+      console.log(color.green("Copied boilerplate"))
+
+      console.log(
+        color.green("Setup done. Now run \"npm install\" to get started")
+      )
     }
-
-    log("Required dependencies successfully installed.")
+    catch (err) {
+      console.error(color.red(err))
+      process.exit(1)
+    }
   }
-)
+
+  main()
+}())